Admin core: Move internal mono logic to new service Application and Template renderer for simplify internal logic complexity.
